+++ lib/whois/record/parser/whois.nic.cz.rb
@@ -15,10 +15,6 @@ module Whois
     class Parser
 
       # Parser for the whois.nic.cz server.
-      #
-      # @note This parser is just a stub and provides only a few basic methods
-      #   to check for domain availability and get domain status.
-      #   Please consider to contribute implementing missing methods.
       # 
       # @see Whois::Record::Parser::Example
       #   The Example parser for the list of all available methods.