+++ spec/whois/record/parser/responses/whois.nic.cz/status_registered_spec.rb
@@ -21,6 +21,16 @@ describe Whois::Record::Parser::WhoisNicCz, "status_registered.expected" do
     @parser = klass.new(part)
   end
 
+  describe "#domain" do
+    it do
+      @parser.domain.should == "google.cz"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      lambda { @parser.domain_id }.should raise_error(Whois::PropertyNotSupported)
+    end
+  end
   describe "#status" do
     it do
       @parser.status.should == :registered
@@ -61,6 +71,68 @@ describe Whois::Record::Parser::WhoisNicCz, "status_registered.expected" do
       @parser.registrar.name.should         == "REG-MARKMONITOR"
       @parser.registrar.organization.should == nil
       @parser.registrar.url.should          == nil
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      @parser.registrant_contacts.should be_a(Array)
+      @parser.registrant_contacts.should have(1).items
+      @parser.registrant_contacts[0].should be_a(Whois::Record::Contact)
+      @parser.registrant_contacts[0].type.should          == Whois::Record::Contact::TYPE_REGISTRANT
+      @parser.registrant_contacts[0].id.should            == "MM12383"
+      @parser.registrant_contacts[0].name.should          == "DNS Admin"
+      @parser.registrant_contacts[0].organization.should  == "Google Inc."
+      @parser.registrant_contacts[0].address.should       == "1600 Amphitheatre Parkway\nMountain View\n94043\nCA\nUS"
+      @parser.registrant_contacts[0].city.should          == nil
+      @parser.registrant_contacts[0].zip.should           == nil
+      @parser.registrant_contacts[0].state.should         == nil
+      @parser.registrant_contacts[0].country.should       == nil
+      @parser.registrant_contacts[0].country_code.should  == nil
+      @parser.registrant_contacts[0].phone.should         == nil
+      @parser.registrant_contacts[0].fax.should           == nil
+      @parser.registrant_contacts[0].email.should         == "dns-admin@google.com"
+      @parser.registrant_contacts[0].created_on.should    == Time.parse("2011-05-18 23:28:26")
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      @parser.admin_contacts.should be_a(Array)
+      @parser.admin_contacts.should have(1).items
+      @parser.admin_contacts[0].should be_a(Whois::Record::Contact)
+      @parser.admin_contacts[0].type.should          == Whois::Record::Contact::TYPE_ADMIN
+      @parser.admin_contacts[0].id.should            == "MM12383"
+      @parser.admin_contacts[0].name.should          == "DNS Admin"
+      @parser.admin_contacts[0].organization.should  == "Google Inc."
+      @parser.admin_contacts[0].address.should       == "1600 Amphitheatre Parkway\nMountain View\n94043\nCA\nUS"
+      @parser.admin_contacts[0].city.should          == nil
+      @parser.admin_contacts[0].zip.should           == nil
+      @parser.admin_contacts[0].state.should         == nil
+      @parser.admin_contacts[0].country.should       == nil
+      @parser.admin_contacts[0].country_code.should  == nil
+      @parser.admin_contacts[0].phone.should         == nil
+      @parser.admin_contacts[0].fax.should           == nil
+      @parser.admin_contacts[0].email.should         == "dns-admin@google.com"
+      @parser.admin_contacts[0].created_on.should    == Time.parse("2011-05-18 23:28:26")
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      @parser.technical_contacts.should be_a(Array)
+      @parser.technical_contacts.should have(1).items
+      @parser.technical_contacts[0].should be_a(Whois::Record::Contact)
+      @parser.technical_contacts[0].id.should            == "MM193020"
+      @parser.technical_contacts[0].name.should          == "Domain Provisioning"
+      @parser.technical_contacts[0].organization.should  == "MarkMonitor, Inc."
+      @parser.technical_contacts[0].address.should       == "10400 Overland Road PMB 155\nBoise\n83709-1433\nID\nUS"
+      @parser.technical_contacts[0].city.should          == nil
+      @parser.technical_contacts[0].zip.should           == nil
+      @parser.technical_contacts[0].state.should         == nil
+      @parser.technical_contacts[0].country.should       == nil
+      @parser.technical_contacts[0].country_code.should  == nil
+      @parser.technical_contacts[0].phone.should         == nil
+      @parser.technical_contacts[0].fax.should           == nil
+      @parser.technical_contacts[0].email.should         == "ccops@markmonitor.com"
+      @parser.technical_contacts[0].created_on.should    == Time.parse("2011-02-03 18:24:34")
     end
   end
   describe "#nameservers" do