+++ spec/whois/record/parser/responses/whois.nic.fo/status_registered_spec.rb
@@ -21,6 +21,16 @@ describe Whois::Record::Parser::WhoisNicFo, "status_registered.expected" do
     @parser = klass.new(part)
   end
 
+  describe "#domain" do
+    it do
+      @parser.domain.should == "nic.fo"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      lambda { @parser.domain_id }.should raise_error(Whois::PropertyNotSupported)
+    end
+  end
   describe "#status" do
     it do
       @parser.status.should == :registered
@@ -57,6 +67,53 @@ describe Whois::Record::Parser::WhoisNicFo, "status_registered.expected" do
   describe "#registrar" do
     it do
       lambda { @parser.registrar }.should raise_error(Whois::PropertyNotSupported)
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      @parser.registrant_contacts.should be_a(Array)
+      @parser.registrant_contacts.should have(1).items
+      @parser.registrant_contacts[0].should be_a(Whois::Record::Contact)
+      @parser.registrant_contacts[0].type.should          == Whois::Record::Contact::TYPE_REGISTRANT
+      @parser.registrant_contacts[0].id.should            == "ID005359"
+      @parser.registrant_contacts[0].name.should          == nil
+      @parser.registrant_contacts[0].organization.should  == "FO-umsitingin"
+      @parser.registrant_contacts[0].address.should       == "Hoydalsvegur 19, Postboks 1255"
+      @parser.registrant_contacts[0].city.should          == "Torshavn"
+      @parser.registrant_contacts[0].zip.should           == "110"
+      @parser.registrant_contacts[0].state.should         == nil
+      @parser.registrant_contacts[0].country.should       == nil
+      @parser.registrant_contacts[0].country_code.should  == "FO" 
+      @parser.registrant_contacts[0].phone.should         == "+298.322400"
+      @parser.registrant_contacts[0].fax.should           == nil
+      @parser.registrant_contacts[0].email.should         == nil
+      @parser.registrant_contacts[0].created_on.should    == Time.parse("2010-07-21 19:11:55")
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      @parser.admin_contacts.should be_a(Array)
+      @parser.admin_contacts.should == []
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      @parser.technical_contacts.should be_a(Array)
+      @parser.technical_contacts.should have(1).items
+      @parser.technical_contacts[0].should be_a(Whois::Record::Contact)
+      @parser.technical_contacts[0].id.should            == "ID005359"
+      @parser.technical_contacts[0].name.should          == nil
+      @parser.technical_contacts[0].organization.should  == "FO-umsitingin"
+      @parser.technical_contacts[0].address.should       == "Hoydalsvegur 19, Postboks 1255"
+      @parser.technical_contacts[0].city.should          == "Torshavn"
+      @parser.technical_contacts[0].zip.should           == "110"
+      @parser.technical_contacts[0].state.should         == nil
+      @parser.technical_contacts[0].country.should       == nil
+      @parser.technical_contacts[0].country_code.should  == "FO" 
+      @parser.technical_contacts[0].phone.should         == "+298.322400"
+      @parser.technical_contacts[0].fax.should           == nil
+      @parser.technical_contacts[0].email.should         == nil
+      @parser.technical_contacts[0].created_on.should    == Time.parse("2010-07-21 19:11:55")
     end
   end
   describe "#nameservers" do